+++ system/src/Grav/Common/Page/Flex/PageIndex.php
@@ -68,7 +68,7 @@ class PageIndex extends FlexPageIndex
      * @param FlexStorageInterface $storage
      * @return array
      */
-    public static function loadEntriesFromStorage(FlexStorageInterface $storage) : array
+    public static function loadEntriesFromStorage(FlexStorageInterface $storage): array
     {
         // Load saved index.
         $index = static::loadIndex($storage);