+++ system/src/Grav/Framework/RequestHandler/Exception/RequestException.php
@@ -78,19 +78,19 @@ class RequestException extends \RuntimeException
     /**
      * @return ServerRequestInterface
      */
-    public function getRequest() : ServerRequestInterface
+    public function getRequest(): ServerRequestInterface
     {
         return $this->request;
     }
 
-    public function getHttpCode() : int
+    public function getHttpCode(): int
     {
         $code = $this->getCode();
 
         return isset(self::$phrases[$code]) ? $code : 500;
     }
 
-    public function getHttpReason() : ?string
+    public function getHttpReason(): ?string
     {
         return self::$phrases[$this->getCode()] ?? self::$phrases[500];
     }