+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 1. [](#new)
     * Added new `-r <job-id>` option for Schduler CLI command to force-run a job [#2720](https://github.com/getgrav/grav/issues/2720)
+1. [](#bugfix)
+    * Regression: Fixed Grav update bug [#2722](https://github.com/getgrav/grav/issues/2722)
 
 # v1.7.0-rc.1
 ## 11/06/2019