+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 1. [](#new)
     * Added new `-r <job-id>` option for Schduler CLI command to force-run a job [#2720](https://github.com/getgrav/grav/issues/2720)
     * Added `Utils::isAssoc()` and `Utils::isNegative()` helper methods
+    * Changed `UserInterface::authorize()` to return `null` having the same meaning as `false` if access is denied because of no matching rule
 1. [](#improved)
     * Improved twig `|array` filter to work with iterators and objects with `toArray()` method
     * Updated Flex `SimpleStorage` code to feature match the other storages